remove unused ambient types
